[FIRRTL][Dedup] Pre-calculate the post-order of modules (#3390)
We modify the instance graph as dedup runs and have some issues with
iterator invalidation. This change copies the module list into a
separate array so we don't have to worry about instance graph updates.

This is not the most efficient solution and in the future we should find
something better.  See https://github.com/llvm/circt/issues/3387 for
more information.