[FIRRTL] InnerSymbolTable: make non-copyable
Copying this class is not currently a correctness issue,
but may be in the future as more functionality is added
(tracking inner symbols, adding/removing them, etc.).
Since copying may be expensive, disallow doing so.

Additionally, the class is meant to be synchronized with an operation
and the inner symbols of its children-- making copies of this
is unlikely to be useful.

Do the same for InnerSymbolCollection.