[FSM] Remove SingleBlockImplicitTerminator from StateOp, TransitionOp
This was essentially enforcing unintended behavior and making it weird to target FSM as a front-end. Previously, `SingleBlockImplicitTerminator<"OutputOp">` required an OutputOp to be in both the output and transition regions, where it really only made sense for the output region.

It's a similar story for the TransitionOp and `SingleBlockImplicitTerminator<"ReturnOp">`.

This, and other small fixes, have been moved to FSMOps.cpp