[FIRRTL] IST: require operations to use our interface.
Ignore operations that happen to have an "inner_sym" attribute,
as it may not necessarily be intended to be interpreted this way.

Nit concern, but also removes another user of the attribute name
outside of the op interface, where we're encapsulating that detail.

Drop the cached attribute for the walk, as well (as a result).
This optimization was borrowed from upstream's SymbolTable and
benchmarks on a real design suggest that it's negligible.
(firtool takes 67.7s before and after, within 0.05s on all runs).