[FIRRTL][GCSM] Use buffer wires for forced output ports
We use buffer wires to break the verilog net when using force
statements.  So far we have only seen a need to do this for input ports,
but the same situation can happen with output ports.  When we are
forcing the value of an output port, we need to insert a wire inside the
module between it and submodule ports, so that the submodule's output
port is not also forced.