[FIRRTL] Remove SubAnnotationAttr
Remove the (now unused) SubAnnotationAttr.  This has been, in prior
commits, fully replaced with a "circt.fieldID" field that is added to
annotations which apply only to part of an aggregate.  This was done
because "circt.fieldID" is viewed as a lighter weight solution that
avoids the need for constant special casing of different "types" of
annotations, i.e., is this a dictionary or is this a SubAnnotationAttr.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>