[FIRRTL] Move all name preservation logic into the drop names pass (#3397)
This change moves more logic related to name preservation into the
DropNames pass.  The parser no longer determines if names should be
droppable at parse time, it always creates wires and registers with
`interesting_name`s.

The drop names pass used to unconditionally mark all operation as
"droppable_name"s when run.  The `DropNames` pass now takes a parameter
if it should preserve everything, nothing, or only things with meaningful
names.

I think that there is a problem with our current naming of "name
preservation": it is not about whether the "name" is preserved or
dropped, it describes whether a wire or register should be preserved or
dropped.  For this reason I think it makes more sense to refer to it as
"value preservation". As the first step toward moving the focus , I
changed the command line option of `firtool` from `-drop-names` to
`--preserve-values=[none | named | all]`.  Alternatives suggestions
welcome!

In the future, I think it would make sense to change the enum attributes
`interesting_name` to `preserved` and `droppable_name` to `droppable`.

This also changes firtool to mark all values droppable by default.