[FIRRTL] Fix GCT Taps for Zero-width
Fix a bug in the Grand Central (GCT) Taps pass where XMRs would be
generated for zero-width data taps.  This then avoids a crash in
LowerToHW where the zero-width, tapped components could _not_ be removed
as they were the targets of XMRs and preventing full lowering of the
FIRRTL IR.  This commit changes the pass to generate no XMRs and this
then enables LowerToHW to correctly remove the zero-width ports on the
data tap module.

Fixes #3458.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>