[FIRRTL] Fix GCT Views for Zero-width
Fix a bug in the Grand Central (GCT) Views pass where a view of a
zero-width component would create an XMR.  Change the GCT View pass to
not create the XMR.  It is fine to leave the zero-width signal in the
interface as ExportVerilog will emit this correctly as a comment.

Fixes #3460.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>