[FIRRTL] Move maybeStringToLocation to utils, NFC
Move a parser utility for converting a string to an optional MLIR
location from the parser into FIRRTLUtils.  This is done because this
method needs to be available to OMIR annotation parsing as OMIR
annotations can contain FIRRTL locations.  The move into a common
utility is done to prevent cyclic dependencies between the transform and
parsing compilation units.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>