[SV] Add sv.logic op (#3455)
This implements sv.logic op which represent "logic" variable in SV spec.
Implementation is almost identical to sv.wire and sv.reg except that
sv.logic is allowed to be defined in procedural and non-procedural
region. If sv.logic is defined in a procedural region, "automatic" keyword is
added to its declaration in emission.