[FIRRTL] Sink Constants into GCT Views XMRs
Change Grand Central (GCT) Views pass to sink trivial constants into
sv::VerbatimOp XMRs inside the mappings module.  This has the benefit of
not needing to add a symbol to the viewed component if it is a constant
which enables this to be deleted and avoids lint warnings related to
unused wires/registers.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>