[SV] Add support for `sv.case` with `hw.enum`
This commit adds support in `sv.case` for switching on enum values. In doing so, a fair bit of refactoring is done to `CasePattern`s to more cleanly handle the different cases of bit, enum, and default patterns.

Since `hw.enum` doesn't yet include any notion of value encoding, the `sv.case` operation exclusively works on either enum or bit pattern.
Furthermore, only a single case value can be compared in a case statement (as opposed to `case A, B : ...`). I think this is a good starting point - if support is needed in the future, it will not be complicated to add.

```mlir
hw.module @AnFSM(%clock : i1) {
  %reg = sv.reg : !hw.inout<!hw.enum<A, B, C>>
  %reg_read = sv.read_inout %reg : !hw.inout<!hw.enum<A, B, C>>

  %A = hw.enum_get #hw.enum.value<A, !hw.enum<A, B, C>>
  %B = hw.enum_get #hw.enum.value<B, !hw.enum<A, B, C>>
  %C = hw.enum_get #hw.enum.value<C, !hw.enum<A, B, C>>

  sv.always posedge %clock {
    sv.case case %reg_read : !hw.enum<A, B, C>
      case A : { sv.passign %reg, %B : !hw.enum<A, B, C> }
      case B : { sv.passign %reg, %C : !hw.enum<A, B, C> }
      default : { sv.passign %reg, %A : !hw.enum<A, B, C> }
  }
}
```

emits
```systemverilog
module AnFSM
    enum {A, B, C} reg_0;
    always @(posedge clock) begin
    case (reg_0)
        A:
        reg_0 <= B;
        B:
        reg_0 <= C;
        default:
        reg_0 <= A;
    endcase
```