[FIRRTL] Fix GCT taps test for constant sink, NFC
Modify existing Grand Central (GCT) Taps tests to work with to-be-added
constant sinking.  For this test, we don't care what the value of some
tapped wire is.  However, we don't want the wire to get a constant value
and then have the GCT Taps pass sink the constant in place of an XMR.
Work around this by creating an invalidated wire with a
DontTouchAnnotation and use that to drive any tapped wires.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>