[FIRRTL] Migrate OMIR parsing to LowerAnnotations (#3466)
Move OMIR annotation scattering from FIRAnnotations.cpp to the
LowerAnnotations pass.  Migrate all existing tests to now use .mlir (as
scattering runs on a pass) and not on .fir.

This require one functional modification to the EmitOMIR pass as SRAM
OMIR NLAs now show up on the module as opposed to on the instance.  The
existing tests pass, but there may be issues with this modification.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>