[docs] Move dialect specific documentation into directories (#3477)
The change moves the dialect specific documentation, such as the many
rationale documents, into the dialect tree. A folder had to be created
for each affected dialect, with an index page containing the orginal
generated contents of the page.

To facilitate this change, I had to update the CIRCT website with some
enhancements made to the MLIR website. I tested this by running the
webpage locally.