[FIRRTL] Add GCT YAML Anno Def, NFC
Add a definition of the Grand Central (GCT) hierarchy file
annotation (which is what is used to control the emission of a YAML
file) and use this in the pass.  This avoids usage of a string which
will be necessarily duplicated inside the LowerAnnotations pass once we
migrate to only using this pass.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>