[HWMemSimImpl] Fix latencies for read-write ports
Fix an issue in the `HWMemSimImpl` pass where latencies would not be
properly modeled for read-write ports. Before this fix, the pass would
pick the larger out of the read and write latency and assemble a delay
pipeline for the read and write portion of the memory with that latency.
This leads to wrong behaviour if the two latencies are different.

This patch instead considers the read and write latency separately. It
takes the minimum of both latencies to construct a common delay pipeline
prefix for the read and write half, and then adds the remaining pipeline
stages to the read and write paths separately.

Added two tests that check for correct behaviour in case of excessive
read and write latencies.