[FIRRTL][SV] Add comment attribute to GC interfaces and modules
We need to add a comment `// VCS coverage exclude_file"` to all Grand
Central modules and interfaces.  When we needed something like this
previousl, it was previously handled by adding a `comment` attribute
onto `hw.module` and handled in lower to HW.

To support these new targets, we have to:

1. Add a comment attribute to SV interfaces.  To accomplish this, I used
   an optional string attribute.  I switched the printed format to print
   the attribute dictionary after the symbol name, which I think is more
   common across all our operations.
2. Update ExportVerilog to print this attribute.  The `emitComment`
   function can recognize null attributes and skip a comment.
3. Propagate a "comment" attribute on FIRRTL modules in LowerToHW to HW
   modules.  This was not added to the ODS arguments, similar to the
   `output_file` attribute.
4. Modify GrandCentral to attach these comments to generated modules and
   interfaces.

I tested this on a design and there were no more files in the `scope`
directory missing this attribute.