[LowerAnnotations] Allow annos without handler to be non-local
If an annotation does not have an explicit handler, still allow it to be
a non-local annotation. Without this, there's a high chance that an
unknown annotation will break firtool, because it might as well contain
a non-local target. Since we can't handle the anno anyway, it seems
reasonable to not further constrain where it applies and leave this to
other passes to figure out (e.g. in a pipeline outside of firtool).