[FSMToSV] Add support for FSM variables
This commit adds support for `fsm.variable` operations in the FSMToSV lowering pass. The meat of this PR is a rewriting of how case muxes are materialized. This procedure is now built around `CaseMuxItem`s which describes a set of assignments to a wire. Furthermore, this rewrite allows for nested case muxes (recursive `CaseMuxItem`s).