[FIRRTL] Move GCT Interface into Companion
Change Grand Central (GCT) Views to move the instantiation of the
interface inside the companion instead of instantiating it in the
parent.  This is done for two reasons.  (1) It simplifies the bindings
file by only needing one bind per View (just the companion) as opposed
to two binds per View (the companion and the interface).  (2) It enables
future changes to use downwards-facing FIRRTL XMRs inside the companion
to drive the interface without having to add ports to the interface.
Reason (2) is viewed as a user-experience benefit as the interface, when
used to populate a waveform, does not include duplicate, extraneous
ports.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>