[LLHD] Remove shifts in favor of comb shifts and dynamic extract operations
Since we have added operations to extract elements and slices from integers, signals, and pointers, and because the Moore compiler recently switched to not using these shifts anymore, we don't need these shift operations anymore to perform a dynamic extract. This also removes the weird corner case where two signals can be concatenated using these operations which are especially difficult to handle in the simulator.