[SystemC] Introduce special types for ports and signals
Remove the current port direction attribute and enum and use types to
represent the port direction. Reuse the PortInfo struct and enum of the
HW dialect to represent the port direction in CPP code when MLIR types
aren't that well suited. The rational is that these constructs are the
same across multiple dialects and should be part of the ModuleLike
interface to be reused by all module ops implementing the HWModuleLike
interface. Providing getters and builders that accept these port
direction/port info struct makes conversion between module ops that
implement HWModuleLike very convenient. This PR also implements that
interface for SystemC modules.
Also switch over to use the module_like and function_like parser and
printer utilities and add support for module visibility and argument
attributes.