[FIRRTL] Use annotations to represent dontTouch
This commit switches back the representation of dontTouch ton normal
annotations. We have used symbols to uniformly represent optimization
blocking behaivor, but it has become hard because of the complexity
hierpath updates or field sensitive symbols. Therefore, this commit
changes the representation of dontTouch to normal annotations.

First, LowerTypes is modified to emit errors for aggregate symbols.
This assumption makes LowerTypes a lot simpler. Actually symbols are
attached before LowerTypes only by dontTouch so the restriction should
not be a big problem. Second, since we don't have to attach symbols at
LowerTypes, the implementation of `filterAnnotation` is simplied a lot.
Also, renmaing of symbols will not happen anymore so they are removed as
well.