Github Actions: automatically build and upload binaries on release  (#3617)
In order to be easily usable with Chisel soon, CIRCT should provide
pre-built binaries at least for macOS and Linux. Currently pre-built
Linux binaries are manually uploaded whenever a new build is made. This
PR adds a Github Actions script that builds and uploads macOS and Linux
binaries whenever a release (or pre-release) is made (requested in
#2842). The build from scratch takes approximately 2 hours.

The generated binaries are not fully static but are portable enough to
run on base versions of macOS and Ubuntu/other Linuxes. The dynamic
linking ends up looking like this:

Linux:

```
$ ldd firtool
	linux-vdso.so.1 (0x00007ffc10fd1000)
	libpthread.so.0 => /lib/x86_64-linux-gnu/libpthread.so.0 (0x00007f19d67c3000)
	libm.so.6 => /lib/x86_64-linux-gnu/libm.so.6 (0x00007f19d6441000)
	libgcc_s.so.1 => /lib/x86_64-linux-gnu/libgcc_s.so.1 (0x00007f19d6229000)
	libc.so.6 => /lib/x86_64-linux-gnu/libc.so.6 (0x00007f19d5e64000)
	/lib64/ld-linux-x86-64.so.2 (0x00007f19d7142000)
```

Mac:

```
$ otool -L firtool
firtool:
	/usr/lib/libz.1.dylib (compatibility version 1.0.0, current version 1.2.11)
	/usr/lib/libSystem.B.dylib (compatibility version 1.0.0, current version 1311.0.0)
```

The build is currently set up to use gcc/g++ instead of clang because
clang includes additional dynamic dependencies on Linux (`librt.so.1`,
`libdl.so.2`, and `libz.so.1`) and does not support the
`-static-libstdc++` flag on mac. The build currently runs on macOS-11
and Ubuntu-20.04.

Currently the uploaded package contains everything in the CIRCT `bin`
directory after a build.