Bump LLVM
Bump LLVM to the current top-of-tree.

This requires changing a few `Attribute`s to the new `TypedAttr`
interface, since attributes no longer carry a type by default. We rely
on this type in quite a few places in the CIRCT codebase, which required
a switch over to `TypedAttr`.

See: https://reviews.llvm.org/D130092