[FIRRTL][IMDCE] Test deleting code w/ref's.  Avoid temp ref wire.
Note: when deleting dead ports, IMDCE introduces temporary wires
for the ports (inside the module) and at points of instantiation.
These are invalid for ref type, which isn't ideal, but they will be erased.

In one case this is not temporary, which this checks for,
the case where an output ref port is alive but not used by instantiators.