[ESI] Fixes for handling none-typed channels in ports. (#3722)
There is basic support for none-typed channels, and a simple test. But
upon using none-typed channels for a more complex Handshake circuit,
it became apparent that the ESI lowering for none-typed ports was
missing a couple special cases. This includes a reduced test case and
the fixes I found necessary.