[FIRRTL][RefOps] Add a pass to lower RefType operations to xmr (#3719)
This commit adds a new pass LowerXMRPass.
The LowerXMRPass replaces every RefResolveOp with a verbatim expr op encoding
the corresponding XMR. The pass also removes every value of RefType and its
uses from the circuit.
This is a dataflow analysis over a very constrained RefType. Domain of the
 dataflow analysis is the set of all RefSendOp. Essentially every RefType
 value must be mapped to one and only one RefSendOp. The analysis propagates
 the reachable RefSendOp to every value of RefType across modules. The
 RefResolveOp is the final leaf into which the dataflow must flow into.
Due to the downward only reference constraint on XMRs, the post order traversal
 ensures that the RefSendOp will be encountered before any user of the
RefType value. The pass fails and errors out if every "live" RefType value is
not mapped to a unique RefSendOp.