[Docs] Update dialects diagram (#3721)
Changing the diagram to update it and improve clarity. Label changes increase clarity of intent. Arrow changes sacrifice *a little* accuracy for visual clarity. Also deleted references to "Circilator" since a google could not find any references aside from a GH issue which implied that it was a theoretical future project.