[PyCDE] Use cocotb-test as test driver (#3725)
No more custom makefile generation and 'make' commands - this is handled by cocotb-test.
Eventually, we expect support for driving cocotb from within python to be included in a future cocotb release. The support in `cocotb-test` is currently identical to the (unreleased) support in upstream `cocotb`. Hence, we use `cocotb-test` since they publish a pypi package that can be easily downloaded.