[FIRParser] Remove more old annotation handling code
Since we switched to using LowerAnnotations exclusively, we no longer
attach annotations to the IR while its being generated.  This lets
us remove the `targetMap` and associated logic for attaching annotations
to FIRRTL entities.

This also changes the "raw annotation" parsing path to append all
annotations to a SmallVector instead of recording them in the
`targetMap` using the "rawAnnotations" key.