[PrepareForEmission] Don't create localparam as a temporary wire (#3770)
Previously if a temporary wire is required for a constant, localparam op is created. However sometimes we truly want to have wires for constants, for example when we want to create wires for sensitivity lists. In such cases the output verilog looks weird because we create an additional temporary wire for localparam. Therefore this PR stops creating a localparam op for constants in the first place.