[FIRRTL][RefOps] LowerXMR: Handle upward references (#3745)
This PR updates the algorithm for lowering RefType Ops to verbatim encoded XMR.
This is required to lower upward scoped references. The dataflow analysis
proceeds in both the forward and backward directions by maintaining the
equivalence classes of the reaching RefSendOp. The dataflow is merged at the
connect inside a parent module. The analysis is based on the constraint that there
is only a single reachable RefSendOp at any RefType val. The analysis also tracks
all the paths between the RefSendOp and the RefResolveOp, to generate the
appropriate XMR path.