[firtool][HW] Add variadic operation splitting
Add a command line option and PrepareForEmission functionality to split
variadic arguments whose operands are composed of many terms.  This is
done to enable large product-of-sum or sum-of-product terms from growing
so large (during inline emission) that they cause poor simulator
performance.  Specifically, this fixes a 10x increase in Verilog
compilation time for Synopsys VCS for an internal design.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>