[ExportChiselInterface] Basic infrastructure for Chisel Interface file emission (#3760)
* [ExportChiselInterface] Basic infrastructure for Chisel Interface file emission

This commit adds a pass for FIRRTL that generates a Scala file with a module
class that represents the interface for the top module of the FIRRTL circuit.
This is to support the development of separable compilation of FIRRTL circuits.
The details about linking circuits together are to be determined. The generated
Scala module extends ExtModule, although a new module class may be introduced
in the future.