[HW] Expose `appendOutputs` in the HWMutableModuleLike interface.
This is a useful piece of generic functionality to be able to add
output values to modules implementing the interface. It is
complementary to the modifyPorts methods that update the module
boundary with new Types. After adding new outputs, it will usually be
necessary to specify which Values should appear on those outputs, and
this additional method makes it possible to do generically. The
default implementation is a no-op, so HWModuleExtern and
HWModuleGenerated will not need to do anything. The implementation on
HWModule overrides this default with the existing behavior.