[SystemC] Add DestructorOp and some helper functions (#3797)
Add an operation to model a C++ destructor that is required to deallocate objects such as a verilated model stored as a pointer that is allocated in the SC_CTOR body.
Also adds a helper function to get the body block more easily and one to the SC_MODULE to get the destructor inside or lazily create one if it does not exist yet. The same helpers are added for the Ctor operation.

Additionally, switch to attr-dict-with-keyword to have the attributes before the region for ctor, func, and cpp.destructor.