[FIRRTL] Add ports to a sequence of InstanceOps via an LCA (#3816)
Add a helper `borePortsOnPath` to add ports along a sequence of `InstanceOps`.
It connects all the ports to each other and forwards a given value from the
source module. `instancePath` specifies a path from a source module through
the `lcaModule` till a target module. Source module is the module referenced
by the first instance on the path  and target module is the referenced target
module of the last instance in the instancePath.

If the `lcaModule` is the source module, then we only need to drill input ports
 starting from the `lcaModule` till the end of `instancePath`. For all other
 cases, we start to drill output ports from the source module  till the
 `lcaModule`, and then input ports from `lcaModule` till the end of
 `instancePath`.

This utility is required for GrandCentral LowerAnnotations to RefOps.
(#3804)