[HandshakeToHW] Don't use rewriter infra in HandshakeToHW
This reverts the "monolithic pattern rewriter" pattern from HandshakeToHW which I've found to be error prone now multiple times (i.e. the HandshakeToHW has a single "top-level" pattern that converts a `handshake.func` to it's HW module, and while doing so, creates a bunch of other modules corresponding to each variation of the Handshake operators). The motivation for this removal has been some local work on elaborating HandshakeToHW, wherein the pattern rewriter (as far as i can tell) tripped up on `BackedgeBuilder` usage when those backedges reside inside a new region created during a rewriter step.