[SystemC] Add dedicated types for integers and bit vectors in SystemC
Merge SystemCTypes.td and SystemCTypesImpl.td in one file as there exist CMake flags to only build types of one dialect now. Thus it allows a downstream dialect to include the file with the definitions directly and we don't need to define a separate constraint for every type. Add types to represent sc_int, sc_uint, sc_bigint, sc_biguint, and sc_bv and an operation to convert between them and the MLIR signless integer type. Adjust the tests to use these new types instead of the builtin integers and also convert to them in the lowering from HW. Support for them in ExportSystemC will be added in a follow up PR.