[FIRRTL][LowerAnnotations] Generate RefOps from GrandCentralView annotations (#3742)
This commit updates the `LowerAnnotations` pass to lower
`GrandCentralViews` to `RefType` connections. The `ref.send`, `ref.resolve` and
`RefType` ports are used to express the cross-module-references for
GrandCentralViews explicitly. The GrandCentral View annotations are now
scattered to the `ref.resolve`, instead of the remote signal. The Interface
generation is still delayed till the GrandCentral pass, which also generates the
verbatim to express the cross-module-reference. 
This is the first step to remove verbatim ops from GrandCentral lowering,
 the next step is to add explicit representation of Interface and
writes-to-Interface in FIRRTL.